LayoutTests/http/tests/security/contentSecurityPolicy/inline-script-blocked-javascript-url-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'options'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'options'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'options'.

CONSOLE MESSAGE: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*".

This test passes if it doesn't alert fail. 
